@@ -3,8 +3,8 @@ package com.google.ar.sceneform;
 import android.content.Context;
 import android.media.Image;
 
-import android.support.annotation.Nullable;
-import android.support.annotation.UiThread;
+import androidx.annotation.Nullable;
+import androidx.annotation.UiThread;
 import android.util.AttributeSet;
 import android.util.Log;
 import android.view.Display;
