@@ -1,7 +1,7 @@
 package com.google.ar.sceneform;
 
-import android.support.annotation.Nullable;
-import android.support.annotation.VisibleForTesting;
+import androidx.annotation.Nullable;
+import androidx.annotation.VisibleForTesting;
 import android.view.MotionEvent;
 import com.google.ar.core.Pose;
 
