@@ -1,8 +1,8 @@
 package com.google.ar.sceneform;
 
 import android.media.Image;
-import android.support.annotation.Nullable;
-import android.support.annotation.VisibleForTesting;
+import androidx.annotation.Nullable;
+import androidx.annotation.VisibleForTesting;
 import android.util.Log;
 import android.view.MotionEvent;
 
