@@ -4,7 +4,7 @@ import android.annotation.SuppressLint;
 import android.content.Context;
 import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 import android.util.Log;
 import android.view.Choreographer;
@@ -332,7 +332,7 @@ public class SceneView extends SurfaceView implements Choreographer.FrameCallbac
 
     if (onBeginFrame(frameTimeNanos)) {
       doUpdate(frameTimeNanos);
-      doRender();
+      doRender(frameTimeNanos);
     }
 
     if (debugEnabled) {
@@ -388,7 +388,7 @@ public class SceneView extends SurfaceView implements Choreographer.FrameCallbac
 
 
 
-  private void doRender() {
+  private void doRender(long frameTimeNanos) {
     Renderer renderer = this.renderer;
     if (renderer == null) {
       return;
@@ -398,7 +398,7 @@ public class SceneView extends SurfaceView implements Choreographer.FrameCallbac
       frameRenderTracker.beginSample();
     }
 
-    renderer.render(debugEnabled);
+    renderer.render(debugEnabled, frameTimeNanos);
 
     if (debugEnabled) {
       frameRenderTracker.endSample();
