@@ -1,7 +1,7 @@
 package com.google.ar.sceneform.rendering;
 
 import android.graphics.SurfaceTexture;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.Surface;
 import com.google.android.filament.Stream;
 import com.google.ar.sceneform.utilities.AndroidPreconditions;
