@@ -1,7 +1,7 @@
 package com.google.ar.sceneform.rendering;
 
 import android.net.Uri;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.Log;
 import com.google.android.filament.IndexBuffer;
 import com.google.android.filament.TextureSampler;
