@@ -1,8 +1,8 @@
 package com.google.ar.sceneform.rendering;
 
 import android.net.Uri;
-import android.support.annotation.Nullable;
-import android.support.annotation.Size;
+import androidx.annotation.Nullable;
+import androidx.annotation.Size;
 import android.util.Log;
 import com.google.android.filament.Engine;
 import com.google.android.filament.Entity;
