@@ -16,10 +16,11 @@ package com.google.ar.sceneform.ux;
 
 import android.app.Activity;
 import android.content.Context;
-import android.support.v7.widget.AppCompatImageView;
 import android.util.AttributeSet;
 import android.view.animation.Animation;
 import android.widget.FrameLayout;
+
+import androidx.appcompat.widget.AppCompatImageView;
 
 
 /** This view contains the hand motion instructions with animation. */
